Moved storage related install/provision to separate directory based on review comments
